Allow manual I/O to bypass filters on arms/conveyors/pumps
The current behavior is the following:
 placing arms/conveyor/pumps with manual I/O toggled on without a filter allows for input/output in the same block face, provided the adjacent block can push or pull.
After placing a filter in in the arm/conveyor/pump, both input and output get affected by what's set on the filter. placing both the input and output in the filter will allow the manual i/o to push or pull, but since the machine can be faster than the adjacent block, it can pull the outputted item, clogging the machine.