Created part one of german translation
Hello to everyone who reads this,

I`m a german guy and I really love this mod, so I decided to make a german translation for the mod.
But I don't have much time so I decided to upload the first part of the translation yet. 

PS: 
I am trying hard to create this config so please be nice to me.
PPS: 
I can also speak and write English, but I use Google translate for things that I don't know what it`s in German.
PPPS:
I will upload the other parts as soon as possible.