New algorithm for Fluid Regulator Keep Exact Mode
Fixes #1705, Supersedes #1710

The existing algorithm was buggy and the proposed rework was hard to
follow the logic, so I came up with a more straightforward algorithm
which should be easier to understand and maintain.

I have also created some tests to demonstrate the functionality of the
revised algorithm and ensure that the behavior is the same as before,
besides the bug.